[BUGFIX] Redirect to Symfony's logout route (#20)
The logout behavior was implemented in a reverse manner. The bundle
assumed that Symfony's logout will first call the OpenID logout handling
and then logout the user from Symfony, which is not the case.

The bundle now expects a dedicated logout route (defaulting to
`logout`), which _may_ need adjustments in the implementing application.

Also, the OpenID configuration is now fetched from Keycloak to obtain
the logout URL in order to avoid hard-coded URLs.

Note: since Symfony 6.4, the `LogoutRouteLoader` may be used, allowing
to configure the logout route to `_logout_main` [1].

[1] https://symfony.com/doc/6.4/security.html#logging-out